requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
boto3==1.34.54
botocore==1.34.54
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
colorama==0.4.6
docx==0.2.4
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.1
flask==3.0.2
Flask-Cors==4.0.0
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
lxml==5.1.0
MarkupSafe==2.1.5
numpy==1.24.4
octoai-sdk==0.9.1
pillow==10.2.0
pycparser==2.21
pydantic==2.6.3
pydantic-core==2.16.3
pypdf2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
requests-file==2.0.0
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
starlette==0.36.3
tldextract==5.1.1
types-PyYAML==6.0.12.12
types-requests==2.31.0.20240218
typing-extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
werkzeug==3.0.1
zipp==3.17.0
zope.event==5.0
zope.interface==6.2
